+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * Animators that are not on the root no longer show a validation warning.  
   Instead they automatically add all things they animate to the exclusions on build.
 * Automatic exclusions are now grouped by their source in the log file.
+* Better selection of root bone and probe anchor when merging meshes with different ones.
+  * Highest priority is for the bone set by the most meshes
+  * Tie breaker is in order of humanoid bone > child of humanoid bone > any other bone
 
 ### Bug Fixes
 * Fix crash when having `Delete Unused Components` disabled.